+++ stats-ring.c
@@ -203,7 +203,7 @@ static void apply_slot(const void *p, void *ctx __unused__)
 
 unsigned int stats_ring_drain(struct stats_ring *ring)
 {
-	uint32_t overflow = 0;
+	uint64_t overflow = 0;
 	uint32_t processed;
 
 	if (ring == NULL)
